attribute directory changes
